Fixture with 1file by entity in progress
